@@ -10,8 +10,8 @@
 </template>
 
 <script>
-export default {
-  name: 'App',
-  components: {}
-}
+  export default {
+    name: 'App',
+    components: {}
+  }
 </script>
