@@ -3,16 +3,16 @@
 </template>
 
 <script>
-export default {
-  name: 'ViewGames',
-  data: () => ({
-    games: []
-  }),
-  mounted() {},
-  methods: {
-    async getGamesByGenre() {
-      // Get Genre Id here
+  export default {
+    name: 'ViewGames',
+    data: () => ({
+      games: []
+    }),
+    mounted() {},
+    methods: {
+      async getGamesByGenre() {
+        // Get Genre Id here
+      }
     }
   }
-}
 </script>
